Version as of Friday, Feb 9 at 11:00 AM PST
This version doesn't work. I'm trying to insert the sliders in the far right via Javascript rather than HTML. This is made necessary by my desire to be able to computationally connect sliders to the global variables. But for some reasons the sliders do not display. I put them in a <div>, preceded and succeeded by labels. The labels show, but not the sliders. I put in a test slider just to insure that my slider parameters work; the testSlider does indeed show. The sliders are successfully appended into the <div>, but they do not appear in the window. 

I have reached the end of my tether; I just cannot see why they aren't showing. The relevant code is lines 208 - 236. What am I doing wrong?